+++ src/Data/Extensible/Instances/Effect.hs
@@ -1,0 +1,9 @@
+{-# OPTIONS_GHC -fno-warn-dodgy-exports #-}
+
+module Data.Extensible.Instances.Effect
+    ( module X
+    ) where
+
+import           Data.Extensible.Effect.BaseIO  as X ()
+import           Data.Extensible.Effect.Logger  as X ()
+import           Data.Extensible.Effect.ThrowIO as X ()