+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Changelog for extensible-instances
 
 ## Unreleased changes
+
+- Write instance
+    - `FromJson`, `ToJson` in `aeson`
+    - `ToRecord`, `ToNamedRecord`, `FromNamedRecord` in `cassava`
+    - `Default` in `data-default`
+    - `MonadBase` in `transformers-base`
+    - `MonadLogger` in `monad-logger`
+    - `MonadThrow` in `exceptions`